Sparse_ann_query ResponsePerSecond(RPS)
