+++ Cargo.toml
@@ -102,3 +102,7 @@ harness = false
 [[bench]]
 name = "sparse_ann_query_benchmark"
 harness = false
+
+[[bench]]
+name = "sparse_ann_query_rps_benchmark"
+harness = false
